requirements.txt
independentsoft.msg